external admin functions
